requirements.txt
python-dotenv
google-genai
transformers
tqdm
numpy
faiss-cpu
sentence-transformers
datasets
scikit-learn
